+++ s1tbx-io/src/main/java/org/esa/s1tbx/io/orbits/doris/DorisOrbitFile.java
@@ -13,9 +13,10 @@
  * You should have received a copy of the GNU General Public License along
  * with this program; if not, see http://www.gnu.org/licenses/
  */
-package org.esa.s1tbx.io.orbits;
+package org.esa.s1tbx.io.orbits.doris;
 
 import com.bc.ceres.core.NullProgressMonitor;
+import org.esa.s1tbx.io.orbits.BaseOrbitFile;
 import org.esa.snap.core.datamodel.MetadataElement;
 import org.esa.snap.core.datamodel.Product;
 import org.esa.snap.core.dataop.downloadable.FtpDownloader;
