+++ s1tbx-io/src/main/java/org/esa/s1tbx/io/orbits/sentinel1/StepAuxdataScraper.java
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License along
  * with this program; if not, see http://www.gnu.org/licenses/
  */
-package org.esa.s1tbx.io.orbits;
+package org.esa.s1tbx.io.orbits.sentinel1;
 
 
 import org.esa.snap.core.util.StringUtils;
