+++ s1tbx-io/src/test/java/org/esa/s1tbx/io/orbits/delft/TestOrbitalDataRecordReader.java
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License along
  * with this program; if not, see http://www.gnu.org/licenses/
  */
-package org.esa.s1tbx.io.orbits;
+package org.esa.s1tbx.io.orbits.delft;
 
 import org.esa.snap.core.util.ResourceInstaller;
 import org.esa.snap.engine_utilities.util.TestUtils;
